backups/stable_version_12_10_25/requirements.txt
python-telegram-bot==21.9
gspread==6.1.4
oauth2client==4.1.3
python-dotenv==1.0.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1